Move Repo struct to common, pass as param to issuetracking client
By passing the repo as a param to the issue tracking client, we are able to stub
it in the issuetracking github tests to avoid being force to use the dev-flow
repo in the test logic. I also moved the Repo struct to the common package to
avoid a circular dependency.